# Conflicts:
#	MiAmusing/MiAmusing.jucer
#	MiAmusing/Source/AmusingScene.cpp
#	MiAmusing/Source/AmusingScene.h
#	MiAmusing/Source/AnimatedPolygon.h
#	MiAmusing/Source/AudioManager.cpp
#	MiAmusing/Source/CompletePolygon.cpp
#	MiAmusing/Source/MultiSceneCanvasManager.cpp
#	MiAmusing/Source/ReadingHead.cpp
#	MiAmusing/Source/editScene.cpp
#	MiAmusing/Source/editScene.h

+++ MiAmusing/Source/AmusingScene.cpp
@@ -369,7 +369,7 @@ void AmusingScene::AddAllIntersections()
 	if (!canvasManagerLocked)
 		throw std::logic_error("Cannot add a new current exciter : cannot get a Unique ID from the canvas manager (not linked to this)");
 
-	int N = areas.size();
+	int N = (int)areas.size();
 	for (int i = 0; i < N; ++i)
 	{
 		for (int j = i + 1; j < N; ++j)
